--- src/main/java/controller/IController.java
@@ -1,4 +1,0 @@
-package controller;
-
-public interface IController {
-}
